+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased][unreleased]
 
+## [0.12.0] - 2023-05-25
 ### Added
 - Class template resolvers allows matching classes by pattern
 - Collect renders also from calls to Latte\Engine
@@ -154,7 +155,8 @@
     - Transform components to explicit calls
 - Error mapper for better DX
 
-[unreleased]: https://github.com/efabrica-team/phpstan-latte/compare/0.11.0...HEAD
+[unreleased]: https://github.com/efabrica-team/phpstan-latte/compare/0.12.0...HEAD
+[0.12.0]: https://github.com/efabrica-team/phpstan-latte/compare/0.11.0...0.12.0
 [0.11.0]: https://github.com/efabrica-team/phpstan-latte/compare/0.10.0...0.11.0
 [0.10.0]: https://github.com/efabrica-team/phpstan-latte/compare/0.9.0...0.10.0
 [0.9.0]: https://github.com/efabrica-team/phpstan-latte/compare/0.8.0...0.9.0